[CHORE] #40 : ManagementCode -> ManagementCodeValidation Rename
