+++ feature/management/src/main/java/com/wap/wapp/feature/management/dialog/ManagementCodeValidationDialog.kt
@@ -1,4 +1,4 @@
-package com.wap.wapp.feature.management.code
+package com.wap.wapp.feature.management.dialog
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -20,13 +20,13 @@ import com.wap.designsystem.WappTheme
 import com.wap.designsystem.component.WappButton
 import com.wap.designsystem.component.WappTextField
 import com.wap.wapp.feature.management.R
-import com.wap.wapp.feature.management.code.ManagementCodeViewModel.ManagementCodeUiState
+import com.wap.wapp.feature.management.dialog.ManagementCodeValidationViewModel.ManagementCodeUiState
 import kotlinx.coroutines.flow.collectLatest
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun ManagementCodeDialog(
-    viewModel: ManagementCodeViewModel = hiltViewModel(),
+fun ManagementCodeValidationDialog(
+    viewModel: ManagementCodeValidationViewModel = hiltViewModel(),
     onDismissRequest: () -> Unit,
     showToast: (Throwable) -> Unit,
 ) {
